# src/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LibraryItemDto": {
				"properties": {
					"userGameId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"launcher": {
						"type": "string"
					},
					"launcherId": {
						"type": "string"
					},
					"coverUrl": {
						"type": "string"
					},
					"gridImageUrl": {
						"type": "string",
						"description": "Absolute local path to the Steam grid portrait image (library_600x900.jpg|png)."
					},
					"description": {
						"type": "string"
					},
					"releaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"genres": {},
					"isInstalled": {
						"type": "boolean"
					},
					"installPath": {
						"type": "string"
					},
					"executablePath": {
						"type": "string"
					},
					"playtime": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"userGameId",
					"id",
					"title",
					"launcher",
					"launcherId",
					"isInstalled",
					"playtime",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"BatchGridImagesResponse": {
				"$ref": "#/components/schemas/Record_string.string_",
				"description": "Map of userGameId → base64 data URL (e.g. \"data:image/jpeg;base64,...\"). Missing images are omitted."
			},
			"BatchGridImagesBody": {
				"properties": {
					"ids": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of userGameIds whose grid images should be returned."
					}
				},
				"required": [
					"ids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LibraryErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameDetailDto": {
				"properties": {
					"userGameId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"launcher": {
						"type": "string"
					},
					"launcherId": {
						"type": "string"
					},
					"coverUrl": {
						"type": "string"
					},
					"gridImageUrl": {
						"type": "string"
					},
					"backgroundImageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"detailedDescription": {
						"type": "string"
					},
					"aboutTheGame": {
						"type": "string"
					},
					"releaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"genres": {},
					"categories": {},
					"screenshots": {},
					"movies": {},
					"developers": {},
					"publishers": {},
					"platforms": {},
					"metacriticScore": {
						"type": "number",
						"format": "double"
					},
					"metacriticUrl": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"supportedLanguages": {
						"type": "string"
					},
					"requiredAge": {
						"type": "number",
						"format": "double"
					},
					"isFree": {
						"type": "boolean"
					},
					"detailsFetchedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"isInstalled": {
						"type": "boolean"
					},
					"installPath": {
						"type": "string"
					},
					"executablePath": {
						"type": "string"
					},
					"playtime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userGameId",
					"id",
					"title",
					"launcher",
					"launcherId",
					"isInstalled",
					"playtime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameActionResult": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"launched": {
						"type": "boolean"
					},
					"protocolUrl": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"launched"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Launcher of Launchers API",
		"version": "1.0.0",
		"description": "Launcher of launchers",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/library/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LibraryItemDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Devuelve todos los juegos de la biblioteca del usuario",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/library/grid-images/batch": {
			"post": {
				"operationId": "GetBatchGridImages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BatchGridImagesResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve las grid images de varios juegos en una sola petición.\r\nHace 1 query a la BD y lee todos los ficheros en paralelo.\r\nLas imágenes ausentes simplemente se omiten del resultado.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BatchGridImagesBody"
							}
						}
					}
				}
			}
		},
		"/library/{userGameId}/icon": {
			"get": {
				"operationId": "GetIcon",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"image/png": {
								"schema": {
									"anyOf": [
										{
											"type": "string",
											"format": "byte"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "Ícono no disponible",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve el ícono PNG embebido en el ejecutable local del juego.\r\n404 cuando el juego no está instalado o no tiene executablePath.\r\nCacheable 24 h — los íconos no cambian entre sincronizaciones.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/library/{userGameId}/grid": {
			"get": {
				"operationId": "GetGridImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"image/*": {
								"schema": {
									"anyOf": [
										{
											"type": "string",
											"format": "byte"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "Grid image no disponible",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve la grid image (portrait) encontrada por el scanner local.\r\nSe sirve por HTTP autenticado para evitar bloqueos del navegador al cargar file://.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/library/{userGameId}/detail": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GameDetailDto"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "Juego no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve el detalle completo de un juego.\r\nSi los datos de la store todavía no han sido obtenidos (detailsFetchedAt IS NULL),\r\nlos obtiene en este momento, los persiste en la DB y los devuelve.\r\nLas llamadas posteriores se sirven directamente desde la cache en DB.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/library/{userGameId}/launch": {
			"post": {
				"operationId": "LaunchGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameActionResult"
								}
							}
						}
					},
					"400": {
						"description": "No se puede lanzar",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Juego no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Lanza un juego instalado.\r\nPrioridad: executablePath guardado → protocol URL del launcher → auto-descubrimiento del .exe.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/library/{userGameId}/install": {
			"post": {
				"operationId": "InstallGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameActionResult"
								}
							}
						}
					},
					"400": {
						"description": "No se puede instalar",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Juego no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Abre el launcher para instalar un juego sincronizado pero no instalado localmente.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/library/{userGameId}/uninstall": {
			"post": {
				"operationId": "UninstallGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameActionResult"
								}
							}
						}
					},
					"404": {
						"description": "Juego no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LibraryErrorResponse"
								}
							}
						}
					}
				},
				"description": "Abre el launcher para desinstalar un juego.\r\nNO borra archivos directamente. El estado se actualiza en la próxima sincronización.",
				"tags": [
					"Library"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userGameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}